+++ pywincffi/core/cdefs/headers/functions.h
@@ -165,6 +165,12 @@ BOOL WINAPI SetHandleInformation(
   _In_ DWORD  dwFlags
 );
 
+// https://msdn.microsoft.com/en-us/aa364992
+DWORD WINAPI GetTempPath(
+  _In_  DWORD  nBufferLength,
+  _Out_ LPTSTR lpBuffer
+);
+
 // https://msdn.microsoft.com/en-us/ms724251
 BOOL WINAPI DuplicateHandle(
   _In_  HANDLE   hSourceProcessHandle,
@@ -207,6 +213,11 @@ HANDLE WINAPI OpenEvent(
 
 // https://msdn.microsoft.com/en-us/ms685081
 BOOL WINAPI ResetEvent(
+  _In_ HANDLE hEvent
+);
+
+// https://msdn.microsoft.com/en-us/library/ms686211
+BOOL WINAPI SetEvent(
   _In_ HANDLE hEvent
 );